Restructure folders for commands and queries
Refs #17
