+++ src/Services/ProfileService/Profile.API/Features/CandidateProfiles/Commands/DeleteCandidate/DeleteCandidateProfileCommand.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace Profile.API.Features.CandidateProfiles.Commands
+namespace Profile.API.Features.CandidateProfiles.Commands.DeleteCandidate
 {
     public record DeleteCandidateProfileCommand(Guid Id) : IRequest<bool>;
 }
