+++ src/Services/ProfileService/Profile.API/Features/CandidateProfiles/Queries/GetCandidateProfile/GetCandidateProfileQuery.cs
@@ -1,7 +1,7 @@
 ﻿using MediatR;
 using Profile.API.DTO;
 
-namespace Profile.API.Features.CandidateProfiles.Queries
+namespace Profile.API.Features.CandidateProfiles.Queries.GetCandidateProfile
 {
     public record GetCandidateProfileQuery(string UserId) : IRequest<CandidateProfileDto?>;
 }
